revert back to cjs
