+++ package.json
@@ -16,7 +16,7 @@
 		"test": "gulp && yarn test-only",
 		"test-ci": "(yarn test-format || 1) && yarn test && yarn install-g && alasql 'select 1 as Succes'",
 		"test-only": "mocha ./test --reporter dot",
-		"test-browser": "node test/browserTestRunner.mjs 7387",
+		"test-browser": "node test/browserTestRunner.js 7387",
 		"test-cover": "# istanbul cover  -x 'lib/zt/zt.js' --dir test/coverage _mocha",
 		"build": "yarn format && gulp",
 		"build-jison": "gulp --jison && gulp",
